Move SDPA to a seperate file
