+++ torchtitan/models/llama/model.py
@@ -8,20 +8,15 @@
 
 
 from dataclasses import dataclass
-from typing import Callable, ClassVar, Optional, Tuple
+from typing import Optional
 
 import torch
-import torch.nn.functional as F
 from torch import nn
-from torch.nn.attention.flex_attention import (
-    BlockMask,
-    create_block_mask,
-    flex_attention,
-)
 
 from torchtitan.components.tokenizer import Tokenizer
 from torchtitan.config_manager import JobConfig
 
+from torchtitan.models.attention import SDPA
 from torchtitan.models.norms import build_norm
 from torchtitan.protocols.train_spec import BaseModelArgs, ModelProtocol
 
@@ -123,7 +118,7 @@ def apply_rotary_emb(
     xq: torch.Tensor,
     xk: torch.Tensor,
     freqs_cis: torch.Tensor,
-) -> Tuple[torch.Tensor, torch.Tensor]:
+) -> tuple[torch.Tensor, torch.Tensor]:
     """
     Apply rotary embeddings to input tensors using the given frequency tensor.
 
@@ -138,7 +133,7 @@ def apply_rotary_emb(
         freqs_cis (torch.Tensor): Precomputed frequency tensor for complex exponentials.
 
     Returns:
-        Tuple[torch.Tensor, torch.Tensor]: Tuple of modified query tensor and key tensor with rotary embeddings.
+        tuple[torch.Tensor, torch.Tensor]: Tuple of modified query tensor and key tensor with rotary embeddings.
     """
     xq_ = torch.view_as_complex(xq.float().reshape(*xq.shape[:-1], -1, 2))
     xk_ = torch.view_as_complex(xk.float().reshape(*xk.shape[:-1], -1, 2))
@@ -179,13 +174,6 @@ class Attention(nn.Module):
 
     """
 
-    # We registered flex_attention related attributes as class variables as we
-    # need to amortize the cost of compilation. Enabling per-instance flex_attention
-    # is not supported.
-    block_mask: ClassVar[Optional[BlockMask]] = None
-    use_flex_attn: ClassVar[bool] = False
-    flex_attn: ClassVar[Optional[Callable]] = None
-
     def __init__(self, model_args: TransformerModelArgs):
         super().__init__()
         self.n_heads = model_args.n_heads
@@ -205,7 +193,7 @@ class Attention(nn.Module):
         self.wo = nn.Linear(
             model_args.n_heads * self.head_dim, model_args.dim, bias=False
         )
-        self.use_flex_attn = model_args.use_flex_attn
+        self.sdpa = SDPA(model_args.use_flex_attn)
 
     def init_weights(self, init_std: float):
         for linear in (self.wq, self.wk, self.wv):
@@ -249,34 +237,13 @@ class Attention(nn.Module):
         xk = keys.transpose(1, 2)  # (bs, n_local_heads, seqlen, head_dim)
         xv = values.transpose(1, 2)  # (bs, n_local_heads, seqlen, head_dim)
 
-        # we use casual mask for training
-        if self.use_flex_attn:
-            # assert False, (type(xq), type(xk), type(xv))
-            self._init_flex_attn(seqlen=seqlen)
-            output = self.flex_attn(xq, xk, xv, block_mask=self.block_mask)
-        else:
-            output = F.scaled_dot_product_attention(xq, xk, xv, is_causal=True)
+        output = self.sdpa(xq, xk, xv)
+
         output = output.transpose(
             1, 2
         ).contiguous()  # (bs, seqlen, n_local_heads, head_dim)
         output = output.view(bs, seqlen, -1)
         return self.wo(output)
-
-    @torch.no_grad()
-    def _init_flex_attn(self, seqlen: int) -> None:
-        if self.block_mask is not None:
-            return
-
-        def causal_mask(b, h, q_idx, kv_idx):
-            return q_idx >= kv_idx
-
-        compiled_create_block_mask = torch.compile(create_block_mask)
-        self.block_mask = compiled_create_block_mask(
-            causal_mask, None, None, seqlen, seqlen
-        )
-        self.flex_attn = torch.compile(
-            flex_attention, mode="max-autotune-no-cudagraphs"
-        )
 
 
 class FeedForward(nn.Module):
